models/TIGER/requirements.txt
accelerated-scan==0.2.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
attrs==24.2.0
audioread==3.0.1
av==13.1.0
blessed==1.20.0
blessings==1.7
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
coloredlogs==15.0.1
conda-pack==0.8.1
ctranslate2==4.5.0
datasets==3.2.0
decorator==5.1.1
dill==0.3.8
docker-pycreds==0.4.0
einops==0.8.0
fast_bss_eval==0.1.4
filelock==3.16.1
flatbuffers==24.3.25
frozenlist==1.5.0
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
gpustat==1.1.1
huggingface-hub==0.26.2
humanfriendly==10.0
hydra-core==1.3.2
HyperPyYAML==1.2.2
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
lazy_loader==0.4
librosa==0.10.2.post1
lightning-utilities==0.11.9
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
modelscope==1.20.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
ninja==1.11.1.3
numba==0.60.0
numpy==2.0.2
omegaconf==2.3.0
onnxruntime==1.20.1
packaging==24.2
pandas==2.2.3
pesq==0.0.4
pillow==11.0.0
pip-system-certs==4.0
platformdirs==4.3.6
pooch==1.8.2
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
ptflops==0.7.4
pyarrow==18.1.0
pycparser==2.22
Pygments==2.18.0
pystoi==0.4.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.0.2
pytorch-ranger==0.1.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rotary-embedding-torch==0.8.5
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentencepiece==0.2.0
sentry-sdk==2.19.0
setproctitle==1.3.4
six==1.16.0
smmap==5.0.1
soundfile==0.12.1
soxr==0.5.0.post1
speechbrain==0.5.14
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.5.1
torch-complex==0.4.4
torch-mir-eval==0.4
torch-optimizer==0.3.0
torchaudio==2.5.1
torchmetrics==1.6.0
torchvision==0.20.1
tqdm==4.67.0
transformers==4.47.1
triton==3.1.0
typeguard==2.13.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wandb==0.18.7
wcwidth==0.2.13
wrapt==1.17.0
xxhash==3.5.0
yarl==1.18.0
